ca-certificates: remove DST_Root_CA_X3 for Let's Encrypt
"DST Root CA X3" expires in 2021-09-30, at which point a bug in GNUTLS
< 3.6.14, OpenSSL 1.0.x, libressl < 3.2.0 causes it to not work with
the Let's Encrypt "android compatibility" chain.

Removing expired certificates from ca-certificates works around the
issue in OpenSSL.

Refer to:
* https://community.letsencrypt.org/t/openssl-client-compatibility-changes-for-let-s-encrypt-certificates/143816
* https://www.agwa.name/blog/post/fixing_the_addtrust_root_expiration

Note: there are other soon-to-expire certificates on ca-certificates.
This patch only concerns itself with the one certificate that is
relevant for Let's Encrypt usage.